:hammer: Refactor: rename promptConfig to questions
