+++ lib/cli.js
@@ -9,7 +9,7 @@ import updateNotifier from 'update-notifier';
 
 import pkg from '../package.json';
 import getConfig from './getConfig';
-import { choices, questions } from './promptConfig';
+import questions, { choices } from './questions';
 
 const configuration = getConfig();
 const choicesList = choices(configuration);
