+++ lib/questions.js
@@ -55,4 +55,5 @@ const questions = (choicesList) => {
   return questionsList;
 };
 
-export { choices, questions };
+export default questions;
+export { choices };
